Auth_Service/pyproject.toml
[tool.poetry]
name = "auth-service"
version = "0.1.0"
description = "This is the Authentication Service Which will provide the authentication services to support the whole services."
authors = ["Iradukunda Fils"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.13.0"
django = "^5.1.4"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"
pillow = "^11.1.0"
djangorestframework-simplejwt = {extras = ["blacklist"], version = "^5.4.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
